added label smoothing

+++ models/feegt.py
@@ -211,7 +211,7 @@ class FEEGT(pl.LightningModule):
         ], device=self.device))
         if self.training and self.use_masking:
             with profiler.record_function("masking"):
-                mask_rand = torch.rand(x.shape[:2], device=self.device)
+                mask_rand = torch.rand(x.shape[:2], dtype=x.dtype, device=self.device)
                 mask = (mask_rand >= self.mask_perc_min) * (mask_rand <= self.mask_perc_max)
                 x[mask] = mask_token
         # adds start and end token
